feat(dpp): renommage fichiers
