+++ front/utils/index.ts
@@ -1,0 +1,5 @@
+import * as EpdsSurveyUtils from './epdsSurvey.util';
+
+export {
+    EpdsSurveyUtils
+};
